Fix error on two complement logic
If value is 0, then !0 will equal 0xffff. When you add one to
that, there's an overflow.

I'm really sure I already fixed this bug for the two_complement
function in the intel 8080 processor. I really need to dedup
that madness.

+++ mos6502cpu/src/bit_utils.rs
@@ -1,6 +1,6 @@
 #[inline]
 pub(crate) fn two_complement(value: u8) -> u8 {
-    !value + 1
+    (!value as u16 + 1) as u8
 }
 
 #[inline]
